Moves Route interface to lib
