+++ src/main/java/ResponseBuilder.java
@@ -1,5 +1,6 @@
 import constants.Codes;
 import lib.Response;
+import lib.Route;
 import routes.*;
 import routes.files.DoggoPNGRoute;
 import routes.files.HealthCheckHTMLRoute;
