+++ src/main/java/lib/Route.java
@@ -1,4 +1,4 @@
-package routes;
+package lib;
 import java.util.ArrayList;
 import java.util.List;
