+++ src/main/java/routes/RedirectRoute.java
@@ -1,6 +1,7 @@
 package routes;
 
 import constants.Headers;
+import lib.Route;
 
 import java.util.ArrayList;
 import java.util.Arrays;
