+++ src/main/java/routes/files/DoggoPNGRoute.java
@@ -1,7 +1,7 @@
 package routes.files;
 
 import constants.Headers;
-import routes.Route;
+import lib.Route;
 
 import java.io.File;
 import java.io.IOException;
